+++ wld/wld_frag.go
@@ -699,14 +699,14 @@ func (e *DMSpriteDef2) ToRaw(wld *Wld, rawWld *raw.Wld) (int16, error) {
 
 	for _, uv := range e.UVs {
 		if wld.WorldDef.NewWorld > 0 {
-			dmSpriteDef.UVs = append(dmSpriteDef.UVs, [2]int32{
-				int32(uv[0]),
-				int32(uv[1]),
+			dmSpriteDef.UVs = append(dmSpriteDef.UVs, [2]float32{
+				float32(uv[0]),
+				float32(uv[1]),
 			})
 		} else {
-			dmSpriteDef.UVs = append(dmSpriteDef.UVs, [2]int32{
-				int32(uv[0] * 256),
-				int32(uv[1] * 256),
+			dmSpriteDef.UVs = append(dmSpriteDef.UVs, [2]float32{
+				float32(uv[0] * 256),
+				float32(uv[1] * 256),
 			})
 		}
 	}
